Tasks Estado Leilao implementada e testada

+++ src/main/java/com/leilaodequadrinhos/api/model/dao/EstadoLeilaoDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface EstadoLeilaoDao {
 
-	EstadoLeilao  findById(Integer id);
+	EstadoLeilao  findById(Long id);
 	List<EstadoLeilao> findAll();
 }
